ci: Fix the condition to run the nox-cross-arch-all job
The current condition makes the job skip unless there is a failure,
which means we also skip on success, which is not right, as we want this
job to succeed if the matrix job succeeds.

To fix this, we only skip if the `nox-cross-arch` matrix job was
skipped and we run otherwise, failing the job if the matrix job was not
successful. We also use the job name explicitly, as we only have one
dependency and it is easier to reason about a string than about arrays.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>