Pass git credentials to the `test-installation` job
Credentials were not being passed to the docker build process, and the
installation test runs as part of the image builting, so it fails when
some `pip` dependency needs git credentials to be fetched.

This commit fixes this by passing the `.git-credentials` file generated
by the `frequenz-floss/gh-action-setup-git` action to the docker build
process.

Even when the produced docker image doesn't leave the GitHub runner,
this is not ideal because we are storing credentials in the docker
image, so eventually this job should be changed so the tests is done
in a running container instead of a docker image building process, but
for now is a good-enough workaround.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>